Add Prometheus middleware to Brig
